+++ lib/src/featureTypes/feature_collection.dart
@@ -13,11 +13,27 @@ class FeatureCollection {
   }
 
   /// Converts the FeatureCollection to a WKT string.
+  ///
+  /// Example:
+  /// ```dart
+  /// FeatureCollection([
+  ///   Point(Coordinate(1, 2)),
+  ///   Point(Coordinate(3, 4))
+  /// ]).toWKT(); // GEOMETRYCOLLECTION(POINT(1 2), POINT(3 4))
+  /// ```
   String toWKT() {
     return 'GEOMETRYCOLLECTION(${features.map((f) => f.toWKT()).join(',')})';
   }
 
   /// Converts the FeatureCollection to a GEOJSON FeatureCollection object.
+  ///
+  /// Example:
+  /// ```dart
+  /// FeatureCollection([
+  ///   Point(Coordinate(1, 2)),
+  ///   Point(Coordinate(3, 4))
+  /// ]).toJson(); // {'type': 'FeatureCollection', 'features': [{'type': 'Feature', 'geometry': {'type': 'Point', 'coordinates': [1, 2]}, 'properties': {}}, {'type': 'Feature', 'geometry': {'type': 'Point', 'coordinates': [3, 4]}, 'properties': {}}]}
+  /// ```
   Map<String, dynamic> toJson() {
     return {
       'type': 'FeatureCollection',
@@ -26,6 +42,11 @@ class FeatureCollection {
   }
 
   /// Creates a FeatureCollection from a JSON object.
+  ///
+  /// Example:
+  /// ```dart
+  /// FeatureCollection.fromJson({'type': 'FeatureCollection', 'features': [{'type': 'Feature', 'geometry': {'type': 'Point', 'coordinates': [1, 2]}, 'properties': {}}, {'type': 'Feature', 'geometry': {'type': 'Point', 'coordinates': [3, 4]}, 'properties': {}}]}); // FeatureCollection([Point(Coordinate(1, 2)), Point(Coordinate(3, 4))])
+  /// ```
   factory FeatureCollection.fromJson(Map<String, dynamic> json) {
     if (json['type'] != 'FeatureCollection') {
       throw ArgumentError('json is not a FeatureCollection');
@@ -53,6 +74,15 @@ class FeatureCollection {
     );
   }
 
+  /// Creates a [FeatureCollection] of [Point]s from the geometries of the [Feature]s contained within.
+  ///
+  /// Example:
+  /// ```dart
+  /// FeatureCollection([
+  ///   LineString([Coordinate(1, 2), Coordinate(3, 4)]),
+  ///   LineString([Coordinate(3, 4), Coordinate(5, 6)])
+  /// ]).explode(); // FeatureCollection([Point(Coordinate(1, 2)), Point(Coordinate(3, 4)), Point(Coordinate(3, 4)), Point(Coordinate(5, 6))])
+  /// ```
   FeatureCollection explode() {
     final explodedFeatures = <Point>[];
     for (final feature in features) {
@@ -67,8 +97,8 @@ class FeatureCollection {
   /// Example:
   /// ```dart
   /// FeatureCollection([
-  ///   Point([Coordinate(1, 2), Coordinate(3, 4)]),
-  ///   Point([Coordinate(5, 6), Coordinate(7, 8)])
+  ///   Point(Coordinate(1, 2)),
+  ///   Point(Coordinate(5, 6))
   /// ]).center; // Point(4, 5)
   /// ```
   Point get center {
@@ -83,5 +113,127 @@ class FeatureCollection {
     }
 
     return Point.fromLatLong(lat / points.length, long / points.length);
+  }
+
+  /// Returns the [BoundingBox] of the [FeatureCollection]
+  /// If the [FeatureCollection] is empty, returns a [BoundingBox.empty]
+  ///
+  /// Example:
+  /// ```dart
+  /// FeatureCollection([
+  ///   LineString([Coordinate(1, 2), Coordinate(3, 4)]),
+  ///   LineString([Coordinate(5, 6), Coordinate(7, 8)])
+  /// ]).bbox; // BoundingBox(1, 2, 7, 8)
+  /// ```
+  BoundingBox get bbox {
+    if (features.isEmpty) {
+      return BoundingBox.empty();
+    }
+
+    List<Point> points = (explode().features as List<Point>);
+    return BoundingBox.fromPoints(points);
+  }
+
+  /// Returns the [BoundingBox]'s [Polygon] of the [FeatureCollection].
+  /// If the [FeatureCollection] is empty, returns a `Polygon([])`.
+  ///
+  /// Example:
+  /// ```dart
+  /// FeatureCollection([
+  ///   LineString([Coordinate(1, 2), Coordinate(3, 4)]),
+  ///   LineString([Coordinate(5, 6), Coordinate(7, 8)])
+  /// ]).envelope;
+  /// ```
+  Polygon get envelope {
+    if (features.isEmpty) {
+      return Polygon([]);
+    }
+
+    List<Point> points = (explode().features as List<Point>);
+    return BoundingBox.fromPoints(points).toPolygon();
+  }
+
+  /// Returns if the [FeatureCollection] is empty.
+  ///
+  /// Example:
+  /// ```dart
+  /// FeatureCollection([
+  ///   LineString([Coordinate(1, 2), Coordinate(3, 4)])
+  /// ]).isEmpty; // false
+  /// FeatureCollection([]).isEmpty; // true
+  /// ```
+  bool get isEmpty => features.isEmpty;
+
+  /// Returns whether all the [Feature]s in the [FeatureCollection] are of the specified [type].
+  /// If the [FeatureCollection] is empty, returns false.
+  /// If the [type] is not a valid geometry type, returns false.
+  ///
+  /// Example:
+  /// ```dart
+  /// FeatureCollection([
+  bool isCollectionOf(String type) {
+    if (![
+      Point.type,
+      MultiPoint.type,
+      LineString.type,
+      MultiLineString.type,
+      Polygon.type,
+      MultiPolygon.type
+    ].contains(type)) {
+      return false;
+    } else if (isEmpty) {
+      return false;
+    }
+
+    switch (type) {
+      case 'Point':
+        return features.every((f) => f is Point);
+      case 'MultiPoint':
+        return features.every((f) => f is MultiPoint);
+      case 'LineString':
+        return features.every((f) => f is LineString);
+      case 'MultiLineString':
+        return features.every((f) => f is MultiLineString);
+      case 'Polygon':
+        return features.every((f) => f is Polygon);
+      case 'MultiPolygon':
+        return features.every((f) => f is MultiPolygon);
+      default:
+        return false;
+    }
+  }
+
+  /// Returns the [Point] with minimum distance between the [FeatureCollection] and the [point].
+  /// If the [FeatureCollection] is empty, returns an empty [FeatureCollection].
+  ///
+  /// Example:
+  /// ```dart
+  /// FeatureCollection([
+  ///   Point([Coordinate(1, 2)]),
+  ///   Point([Coordinate(5, 6)])
+  /// ]).nearest(Point([Coordinate(0, 0)])); // Point(1, 2)
+  /// FeatureCollection([]).nearest(Point([Coordinate(0, 0)])); // null
+  /// FeatureCollection([
+  ///  LineString([Coordinate(1, 2), Coordinate(3, 4)]),
+  /// ```
+  Point? nearestPointTo(Point point) {
+    if (isEmpty) {
+      return null;
+    } else if (!isCollectionOf(Point.type)) {
+      return null;
+    }
+
+    double minDistance = double.infinity;
+    Point nearestPoint = features[0] as Point;
+
+    for (final feature in (features as List<Point>)) {
+      final distance = feature.coordinate.distanceTo(point.coordinate);
+      if (distance < minDistance) {
+        minDistance = distance;
+        nearestPoint = feature;
+      }
+    }
+
+    return nearestPoint;
   }
 }