+++ lib/src/featureTypes/line_string.dart
@@ -80,6 +80,17 @@ class LineString extends Feature {
     );
   }
 
+  /// Creates a [LineString] with random [Coordinate]s
+  /// Currently only supports a 2 [Coordinate] [LineString]
+  ///
+  /// Example:
+  /// ```dart
+  /// LineString.random(); // LineString([Coordinate(?, ?), Coordinate(?, ?)])
+  /// ```
+  factory LineString.random() {
+    return LineString([Coordinate.random(), Coordinate.random()]);
+  }
+
   /// Explodes the [LineString] into a [List] of [Point]s
   ///
   /// Example:
@@ -116,6 +127,23 @@ class LineString extends Feature {
   BoundingBox get bbox {
     List<Point> points = explode();
     return BoundingBox.fromPoints(points);
+  }
+
+  /// Returns each segment (a 2-coordinate [LineString]) of the [LineString] in a [FeatureCollection]
+  ///
+  /// Example:
+  /// ```dart
+  /// LineString([Coordinate(1, 2), Coordinate(3, 4), Coordinate(5, 6)]).segments; // FeatureCollection([LineString([Coordinate(1, 2), Coordinate(3, 4)]), LineString([Coordinate(3, 4), Coordinate(5, 6)])])
+  /// ```
+  FeatureCollection get segments {
+    List<LineString> segments = [];
+    for (int i = 0; i < coordinates.length - 1; i++) {
+      final start = coordinates[i];
+      final end = coordinates[i + 1];
+      final segment = LineString([start, end]);
+      segments.add(segment);
+    }
+    return FeatureCollection(segments);
   }
 
   /// If the [LineString] is a closed ring, it will be converted to a [Polygon].